ui/volumes: Extract health comparison logic to utils file to use it both on initial data sorting and react-table sorting.
Refs: #2876
